feature/ASAA-18 Moved BuildConfigProvider to domain and set PlacesModule with Map Key
