+++ app/src/main/java/com/bottlerocketstudios/mapsdemo/di/AppModule.kt
@@ -1,7 +1,7 @@
 package com.bottlerocketstudios.mapsdemo.di
 
 import com.bottlerocketstudios.mapsdemo.buildconfig.BuildConfigProviderImpl
-import com.bottlerocketstudios.mapsdemo.data.di.BuildConfigProvider
+import com.bottlerocketstudios.mapsdemo.domain.di.BuildConfigProvider
 import org.koin.dsl.module
 
 object AppModule {
