+++ places/src/main/java/com/bottlerocketstudios/places/startup/PlacesStartupInitializer.kt
@@ -2,12 +2,16 @@ package com.bottlerocketstudios.places.startup
 
 import android.content.Context
 import androidx.startup.Initializer
+import com.bottlerocketstudios.mapsdemo.domain.di.BuildConfigProvider
 import com.google.android.libraries.places.api.Places
 import org.koin.android.BuildConfig
+import org.koin.java.KoinJavaComponent.inject
+
 
 class PlacesStartupInitializer: Initializer<Unit>{
     override fun create(context: Context) {
-        Places.initialize(context, BuildConfig.MAPS_API_KEY)
+        val buildConfigProvider by inject<BuildConfigProvider>(BuildConfigProvider::class.java)
+        Places.initialize(context, buildConfigProvider.mapsApiKey)
     }
 
     override fun dependencies(): List<Class<out Initializer<*>>> = emptyList()
